fix: change type int to double for divide

+++ calculator/Program.cs
@@ -6,13 +6,13 @@
         {
             Console.WriteLine("Hesap Makinası");
             Console.Write("BİRİNCİ SAYIYI GİRİNİZ: ");
-            int sayi1=int.Parse(Console.ReadLine());
+            double sayi1=double.Parse(Console.ReadLine());
 
             Console.Write("Lütfen bir işlem seçiniz(+) (-) (*) (/) (%) :  ");
             string islem = Console.ReadLine();
 
             Console.Write("ikinci sayıyı giriniz: ");
-            int sayi2=int.Parse(Console.ReadLine());
+            double sayi2=double.Parse(Console.ReadLine());
 
 
             switch (islem)
@@ -30,6 +30,7 @@
                 case "/":
                     Console.WriteLine(sayi1 / sayi2);
                     break;
+               
             }
         }
     }
